+++ tests/resolveConfig.test.js
@@ -1763,3 +1763,30 @@ test('all helpers can be destructured from the first function argument', () => {
     },
   })
 })
+
+test('does not duplicate extended configs every time resolveConfig is called', () => {
+  let shared = {
+    foo: { bar: { baz: [{ color: 'red' }] } },
+  }
+
+  const createConfig = (color) =>
+    resolveConfig([
+      {
+        theme: {
+          foo: shared.foo,
+          extend: {
+            foo: { bar: { baz: { color } } },
+          },
+        },
+      },
+    ])
+
+  createConfig('orange')
+  createConfig('yellow')
+  createConfig('green')
+
+  const result = createConfig('blue')
+
+  expect(shared.foo.bar.baz).toMatchObject([{ color: 'red' }])
+  expect(result.theme.foo.bar.baz).toMatchObject([{ color: 'red' }, { color: 'blue' }])
+})